+++ kmath-ast/src/commonMain/kotlin/space/kscience/kmath/ast/rendering/SyntaxRenderer.kt
@@ -5,12 +5,15 @@
 
 package space.kscience.kmath.ast.rendering
 
+import space.kscience.kmath.misc.UnstableKMathAPI
+
 /**
  * Abstraction of writing [MathSyntax] as a string of an actual markup language. Typical implementation should
  * involve traversal of MathSyntax with handling each its subtype.
  *
  * @author Iaroslav Postovalov
  */
+@UnstableKMathAPI
 public fun interface SyntaxRenderer {
     /**
      * Renders the [MathSyntax] to [output].
@@ -23,6 +26,7 @@ public fun interface SyntaxRenderer {
  *
  * @author Iaroslav Postovalov
  */
+@UnstableKMathAPI
 public fun SyntaxRenderer.renderWithStringBuilder(node: MathSyntax): String {
     val sb = StringBuilder()
     render(node, sb)