+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/Structure1D.kt
@@ -6,6 +6,8 @@
 package space.kscience.kmath.nd
 
 import space.kscience.kmath.structures.Buffer
+import space.kscience.kmath.structures.MutableBuffer
+import space.kscience.kmath.structures.asMutableBuffer
 import space.kscience.kmath.structures.asSequence
 import kotlin.jvm.JvmInline
 
@@ -26,6 +28,16 @@ public interface Structure1D<T> : StructureND<T>, Buffer<T> {
 }
 
 /**
+ * A mutable structure that is guaranteed to be one-dimensional
+ */
+public interface MutableStructure1D<T> : Structure1D<T>, MutableStructureND<T>, MutableBuffer<T> {
+    public override operator fun set(index: IntArray, value: T) {
+        require(index.size == 1) { "Index dimension mismatch. Expected 1 but found ${index.size}" }
+        set(index[0], value)
+    }
+}
+
+/**
  * A 1D wrapper for nd-structure
  */
 @JvmInline
@@ -35,6 +47,23 @@ private value class Structure1DWrapper<T>(val structure: StructureND<T>) : Struc
 
     override operator fun get(index: Int): T = structure[index]
     override fun elements(): Sequence<Pair<IntArray, T>> = structure.elements()
+}
+
+/**
+ * A 1D wrapper for a mutable nd-structure
+ */
+private class MutableStructure1DWrapper<T>(val structure: MutableStructureND<T>) : MutableStructure1D<T> {
+    override val shape: IntArray get() = structure.shape
+    override val size: Int get() = structure.shape[0]
+    override fun elements(): Sequence<Pair<IntArray, T>> = structure.elements()
+
+    override fun get(index: Int): T = structure[index]
+    override fun set(index: Int, value: T) {
+        structure[intArrayOf(index)] = value
+    }
+
+    override fun copy(): MutableBuffer<T> =
+        structure.elements().map { it.second }.toMutableList().asMutableBuffer()
 }
 
 
@@ -52,6 +81,21 @@ private value class Buffer1DWrapper<T>(val buffer: Buffer<T>) : Structure1D<T> {
     override operator fun get(index: Int): T = buffer[index]
 }
 
+internal class MutableBuffer1DWrapper<T>(val buffer: MutableBuffer<T>) : MutableStructure1D<T> {
+    override val shape: IntArray get() = intArrayOf(buffer.size)
+    override val size: Int get() = buffer.size
+
+    override fun elements(): Sequence<Pair<IntArray, T>> =
+        buffer.asSequence().mapIndexed { index, value -> intArrayOf(index) to value }
+
+    override operator fun get(index: Int): T = buffer[index]
+    override fun set(index: Int, value: T) {
+        buffer[index] = value
+    }
+
+    override fun copy(): MutableBuffer<T> = buffer.copy()
+}
+
 /**
  * Represent a [StructureND] as [Structure1D]. Throw error in case of dimension mismatch
  */
@@ -61,6 +105,11 @@ public fun <T> StructureND<T>.as1D(): Structure1D<T> = this as? Structure1D<T> ?
         else -> Structure1DWrapper(this)
     }
 } else error("Can't create 1d-structure from ${shape.size}d-structure")
+
+public fun <T> MutableStructureND<T>.as1D(): MutableStructure1D<T> =
+    this as? MutableStructure1D<T> ?: if (shape.size == 1) {
+        MutableStructure1DWrapper(this)
+    } else error("Can't create 1d-structure from ${shape.size}d-structure")
 
 /**
  * Represent this buffer as 1D structure
@@ -75,3 +124,4 @@ internal fun <T : Any> Structure1D<T>.unwrap(): Buffer<T> = when {
     this is Structure1DWrapper && structure is BufferND<T> -> structure.buffer
     else -> this
 }
+