+++ kmath-tensors/src/commonMain/kotlin/space/kscience/kmath/tensors/api/Tensor.kt
@@ -1,0 +1,5 @@
+package space.kscience.kmath.tensors.api
+
+import space.kscience.kmath.nd.MutableStructureND
+
+public typealias Tensor<T> = MutableStructureND<T>